requirements.txt
bleach==1.5.0
decorator==4.1.2
enum34==1.1.6
h5py==2.7.1
html5lib==0.9999999
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
Keras==2.1.2
Markdown==2.6.9
numpy==1.13.3
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
protobuf==3.4.0
ptyprocess==0.5.2
Pygments==2.2.0
PyYAML==3.12
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0rc2
tqdm==4.19.4
traitlets==4.3.2
typing==3.6.2
wcwidth==0.1.7
Werkzeug==0.12.2
sgfmill==1.0.0
pygtp==0.4
